+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Linting now recognised `shell` blocks to avoid error `when: condition has too many lines` ([#1557](https://github.com/nf-core/tools/issues/1557))
 - Linting: fix error when using comments after `input` tuple lines ([#1542](https://github.com/nf-core/tools/issues/1542))
 - Linting: Don't lint the `shell` block when `script` is used ([1558](https://github.com/nf-core/tools/pull/1558))
+- Linting: Check that `template` is used in `script` blocks
 
 ## [v2.3.2 - Mercury Vulture Fixed Formatting](https://github.com/nf-core/tools/releases/tag/2.3.2) - [2022-03-24]